Moved like helper methods to LikeController, and added like controller to location constructor, adds tests to cover the like helper methods for 100% coverage B)
